pyproject.toml
[project]
name = "minimalagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.1",
    "aiortc>=1.6.0",
    "asyncpg>=0.29.0",
    "dotenv>=0.9.9",
    "fastapi>=0.104.1",
    "fuzzywuzzy>=0.18.0",
    "google-cloud-speech>=2.33.0",
    "google-cloud-texttospeech>=2.30.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.25.2",
    "loguru>=0.7.0",
    "nltk>=3.8.1",
    "numpy>=1.24.4",
    "pillow>=10.1.0",
    "pipecat-ai[cartesia,daily,deepgram,local-smart-turn-v3,silero,webrtc,whisper]>=0.0.85",
    "pipecat-ai-flows>=0.0.21",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-levenshtein>=0.23.0",
    "redis>=5.0.1",
    "scikit-learn>=1.3.2",
    "structlog>=23.2.0",
    "supabase>=2.6.0",
    "textblob>=0.17.1",
    "uvicorn[standard]>=0.24.0",
    "simli-ai>=0.1.19",
    "pipecat-ai-small-webrtc-prebuilt>=1.0.0",
    "openai-whisper>=20250625",
    "google-genai>=1.41.0",
]
